Tiro não alinhado com o player no zoom da camera

+++ GameEngineSolution/GameEngine/InputManager.h
@@ -23,8 +23,8 @@ class InputManager {
 		int GetMouseX();
 		int GetMouseY();
 
-		int GetWorldMouseX();
-		int GetWorldMouseY();
+		float GetWorldMouseX();
+		float GetWorldMouseY();
 
 		bool QuitRequested();
 		static InputManager& GetInstance();
